Move service registration in System Bridge integration to async_setup
